+++ src/plans/stopping_criterion.jl
@@ -212,19 +212,6 @@ function StopWhenChangeLess(
         ε, "", storage, inverse_retraction_method, 0
     )
 end
-function StopWhenChangeLess(
-    ε::Float64;
-    storage::StoreStateAction=StoreStateAction([:Iterate]),
-    manifold::AbstractManifold=DefaultManifold(),
-    inverse_retraction_method::IRT=default_inverse_retraction_method(manifold),
-) where {IRT<:AbstractInverseRetractionMethod}
-    if !(manifold isa DefaultManifold)
-        @warn "The `manifold` keyword is deprecated, use the first positional argument `M` instead."
-    end
-    return StopWhenChangeLess{IRT,typeof(storage)}(
-        ε, "", storage, inverse_retraction_method, 0
-    )
-end
 function (c::StopWhenChangeLess)(mp::AbstractManoptProblem, s::AbstractManoptSolverState, i)
     if i == 0 # reset on init
         c.reason = ""
@@ -866,13 +853,13 @@ If either `s1` (or `s2`) is already an [`StopWhenAll`](@ref), then `s2` (or `s1`
 appended to the list of [`StoppingCriterion`](@ref) within `s1` (or `s2`).
 
 # Example
-    a = StopAfterIteration(200) & StopWhenChangeLess(1e-6)
+    a = StopAfterIteration(200) & StopWhenChangeLess(M, 1e-6)
     b = a & StopWhenGradientNormLess(1e-6)
 
 Is the same as
 
-    a = StopWhenAll(StopAfterIteration(200), StopWhenChangeLess(1e-6))
-    b = StopWhenAll(StopAfterIteration(200), StopWhenChangeLess(1e-6), StopWhenGradientNormLess(1e-6))
+    a = StopWhenAll(StopAfterIteration(200), StopWhenChangeLess(M, 1e-6))
+    b = StopWhenAll(StopAfterIteration(200), StopWhenChangeLess(M, 1e-6), StopWhenGradientNormLess(1e-6))
 """
 function Base.:&(s1::S, s2::T) where {S<:StoppingCriterion,T<:StoppingCriterion}
     return StopWhenAll(s1, s2)
@@ -955,13 +942,13 @@ If either `s1` (or `s2`) is already an [`StopWhenAny`](@ref), then `s2` (or `s1`
 appended to the list of [`StoppingCriterion`](@ref) within `s1` (or `s2`)
 
 # Example
-    a = StopAfterIteration(200) | StopWhenChangeLess(1e-6)
+    a = StopAfterIteration(200) | StopWhenChangeLess(M, 1e-6)
     b = a | StopWhenGradientNormLess(1e-6)
 
 Is the same as
 
-    a = StopWhenAny(StopAfterIteration(200), StopWhenChangeLess(1e-6))
-    b = StopWhenAny(StopAfterIteration(200), StopWhenChangeLess(1e-6), StopWhenGradientNormLess(1e-6))
+    a = StopWhenAny(StopAfterIteration(200), StopWhenChangeLess(M, 1e-6))
+    b = StopWhenAny(StopAfterIteration(200), StopWhenChangeLess(M, 1e-6), StopWhenGradientNormLess(1e-6))
 """
 function Base.:|(s1::S, s2::T) where {S<:StoppingCriterion,T<:StoppingCriterion}
     return StopWhenAny(s1, s2)