+++ src/solvers/quasi_Newton.jl
@@ -172,7 +172,7 @@ function default_stepsize(
         M;
         retraction_method=retraction_method,
         vector_transport_method=vector_transport_method,
-        linesearch_stopsize=1e-10,
+        stop_when_stepsize_less=1e-10,
     )
 end
 @doc raw"""